feat(search): several improvements
  - show number of hidden posts because of blacklist
  - "post publically" shows poster
  - "post publically" button disabled after posting
  - check search tags for blacklisted tags early and fail
  - error handling on booru
  - disable random: true on order/sort tags
  - check blacklisted types
  - check blacklisted ratings
